Define GPT contract
